+++ toolbox/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.1.6]
+- Adding additional samples 
+
 ## [0.1.5]
 - Add support for running exported brains
 
@@ -16,7 +19,7 @@ All notable changes to this project will be documented in this file.
 ## [0.1.2]
 
 - #16618: Stop requesting new training events on unregister.
-- #16619: Update Mathworks cartpole sample license.
+- #16619: Update MathWorks cartpole sample license.
 
 ## [0.1.1]